Docs/last stretch

+++ contracts/v2/structs/JBSplit.sol
@@ -7,8 +7,8 @@ struct JBSplit {
   // A flag that only has effect if a projectId is also specified, and that project has issued its tokens.
   // If so, this flag indicates if the tokens that result from making a payment to the project should be delivered staked or unstaked to the beneficiary.
   bool preferClaimed;
-  // The percent of the whole group that this split occupies. This number is out of 10000.
-  uint16 percent;
+  // The percent of the whole group that this split occupies. This number is out of 10000000.
+  uint24 percent;
   // Specifies if the split should be unchangeable until the specifies time comes, with the exception of extending the lockedUntil period.
   uint48 lockedUntil;
   // The role the  beneficary depends on whether or not projectId is specified, or whether or not allocator is specified.
